separate newspan and continuespan interceptor
