+++ tracing-opentelemetry/src/main/java/io/micronaut/tracing/opentelemetry/interceptor/AbstractOpenTelemetryTraceInterceptor.java
@@ -1,0 +1,70 @@
+/*
+ * Copyright 2017-2022 original authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.micronaut.tracing.opentelemetry.interceptor;
+
+import io.micronaut.aop.InterceptPhase;
+import io.micronaut.aop.MethodInterceptor;
+import io.micronaut.aop.MethodInvocationContext;
+import io.micronaut.core.annotation.AnnotationMetadata;
+import io.micronaut.core.type.Argument;
+import io.micronaut.tracing.annotation.SpanTag;
+import io.opentelemetry.api.trace.Span;
+import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
+import io.opentelemetry.instrumentation.api.util.ClassAndMethod;
+import jakarta.inject.Named;
+
+/**
+ * Implements tracing logic for <code>ContinueSpan</code> and <code>NewSpan</code>
+ * using the Open Telemetry API.
+ *
+ * @author Nemanja Mikic
+ * @since 4.1.0
+ */
+public abstract class AbstractOpenTelemetryTraceInterceptor implements MethodInterceptor<Object, Object> {
+
+    protected final Instrumenter<ClassAndMethod, Object> instrumenter;
+
+    /**
+     * Initialize the interceptor with tracer and conversion service.
+     *
+     * @param instrumenter the ClassAndMethod Instrumenter
+     */
+    protected AbstractOpenTelemetryTraceInterceptor(@Named("micronautCodeTelemetryInstrumenter") Instrumenter<ClassAndMethod, Object> instrumenter) {
+        this.instrumenter = instrumenter;
+    }
+
+    @Override
+    public int getOrder() {
+        return InterceptPhase.TRACE.getPosition();
+    }
+
+    public static void tagArguments(MethodInvocationContext<?, ?> context) {
+        Argument<?>[] arguments = context.getArguments();
+        Object[] parameterValues = context.getParameterValues();
+        for (int i = 0; i < arguments.length; i++) {
+            Argument<?> argument = arguments[i];
+            AnnotationMetadata annotationMetadata = argument.getAnnotationMetadata();
+            if (annotationMetadata.hasAnnotation(SpanTag.class)) {
+                Object v = parameterValues[i];
+                if (v != null) {
+                    String tagName = annotationMetadata.stringValue(SpanTag.class).orElse(argument.getName());
+                    Span span = Span.current();
+                    span.setAttribute(tagName, v.toString());
+                }
+            }
+        }
+    }
+}
