+++ tracing-opentelemetry/src/test/groovy/io/micronaut/tracing/instrument/util/AnnotationMappingSpec.groovy
@@ -4,11 +4,7 @@ import groovy.util.logging.Slf4j
 import io.micronaut.context.ApplicationContext
 import io.micronaut.core.annotation.Introspected
 import io.micronaut.http.HttpRequest
-import io.micronaut.http.annotation.Body
-import io.micronaut.http.annotation.Controller
-import io.micronaut.http.annotation.Get
-import io.micronaut.http.annotation.Header
-import io.micronaut.http.annotation.Post
+import io.micronaut.http.annotation.*
 import io.micronaut.http.client.annotation.Client
 import io.micronaut.reactor.http.client.ReactorHttpClient
 import io.micronaut.runtime.server.EmbeddedServer
@@ -16,6 +12,7 @@ import io.micronaut.scheduling.annotation.ExecuteOn
 import io.micronaut.tracing.annotation.ContinueSpan
 import io.micronaut.tracing.annotation.NewSpan
 import io.micronaut.tracing.annotation.SpanTag
+import io.opentelemetry.api.common.AttributeKey
 import io.opentelemetry.extension.annotations.SpanAttribute
 import io.opentelemetry.extension.annotations.WithSpan
 import io.opentelemetry.sdk.testing.exporter.InMemorySpanExporter
@@ -27,6 +24,7 @@ import reactor.util.function.Tuples
 import spock.lang.AutoCleanup
 import spock.lang.Shared
 import spock.lang.Specification
+import spock.util.concurrent.PollingConditions
 
 import static io.micronaut.scheduling.TaskExecutors.IO
 
@@ -42,6 +40,8 @@ class AnnotationMappingSpec extends Specification {
     @Shared
     @AutoCleanup
     ReactorHttpClient reactorHttpClient = ReactorHttpClient.create(embeddedServer.URL)
+
+    private PollingConditions conditions = new PollingConditions()
 
     void 'test map WithSpan annotation'() {
         int count = 1
@@ -68,7 +68,7 @@ class AnnotationMappingSpec extends Specification {
 
         testExporter.finishedSpanItems.size() == count * spanNumbers
 
-        !testExporter.finishedSpanItems.attributes.any(x -> x.asMap().keySet().any(y -> y.key == "tracing-annotation-span-attribute"))
+        testExporter.finishedSpanItems.attributes.any(x -> x.asMap().keySet().any(y -> y.key == "tracing-annotation-span-attribute"))
         !testExporter.finishedSpanItems.attributes.any(x -> x.asMap().keySet().any(y -> y.key == "tracing-annotation-span-tag-no-withspan"))
         testExporter.finishedSpanItems.attributes.any(x -> x.asMap().keySet().any(y -> y.key == "tracing-annotation-span-tag-with-withspan"))
         testExporter.finishedSpanItems.attributes.any(x -> x.asMap().keySet().any(y -> y.key == "tracing-annotation-span-tag-continue-span"))
@@ -79,6 +79,25 @@ class AnnotationMappingSpec extends Specification {
         cleanup:
         testExporter.reset()
     }
+
+    void 'client with tracing annotations' () {
+        def testExporter = embeddedServer.applicationContext.getBean(InMemorySpanExporter)
+        def warehouseClient = embeddedServer.applicationContext.getBean(WarehouseClient)
+
+        expect:
+
+        warehouseClient.order(Map.of("testOrderKey","testOrderValue"))
+        warehouseClient.getItemCount("testItemCount", 10) == 10
+        conditions.eventually {
+            testExporter.finishedSpanItems.size() == 1
+            testExporter.finishedSpanItems.get(0).name == "WarehouseClient.order"
+            testExporter.finishedSpanItems.get(0).attributes.get(AttributeKey.stringKey("warehouse.order")) == "{testOrderKey=testOrderValue}"
+        }
+
+        cleanup:
+        testExporter.reset()
+    }
+
 
     @Introspected
     static class SomeBody {
@@ -102,6 +121,7 @@ class AnnotationMappingSpec extends Specification {
 
         @ExecuteOn(IO)
         @Get("/test")
+        @ContinueSpan
         Mono<String> test(@SpanAttribute("tracing-annotation-span-attribute")
                           @Header("X-TrackingId") String tracingId) {
             LOG.debug("test")
@@ -129,5 +149,34 @@ class AnnotationMappingSpec extends Specification {
         String methodContinueSpan(@SpanTag("tracing-annotation-span-tag-continue-span") String tracingId) {
             return tracingId
         }
+    }
+
+    @Controller("/client")
+    static class ClientController {
+
+        @Get("/count")
+        int getItemCount(@QueryValue String store, @SpanTag @QueryValue int upc) {
+            return upc
+        }
+
+
+        @Post("/order")
+        void order(@SpanTag("warehouse.order") Map<String, ?> json) {
+
+        }
+
+    }
+
+    @Client("/client")
+    static interface WarehouseClient {
+
+        @Get("/count")
+        @ContinueSpan
+        int getItemCount(@QueryValue String store, @SpanTag @QueryValue int upc);
+
+        @Post("/order")
+        @NewSpan
+        void order(@SpanTag("warehouse.order") Map<String, ?> json);
+
     }
 }
